# server/qa-baselines/openapi.current.json
{
  "components": {
    "schemas": {
      "ErrorEnvelope": {
        "additionalProperties": false,
        "description": "Unified error envelope: every API failure renders as {error:{code,message,details?}}.",
        "properties": {
          "error": {
            "additionalProperties": false,
            "properties": {
              "code": {
                "description": "Stable code from the ERROR_CODES catalog.",
                "type": "string"
              },
              "details": {
                "additionalProperties": true,
                "type": "object"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "cookieAuth": {
        "in": "cookie",
        "name": "novel_engine_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Self-hosted writing studio API (TypeScript rewrite).",
    "title": "Novel Engine API",
    "version": "0.6.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/projects/{projectId}/exports": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "exports": {
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "checksum_sha256": {
                            "pattern": "^[a-f0-9]{64}$",
                            "type": "string"
                          },
                          "created_at": {
                            "format": "date-time",
                            "type": "string"
                          },
                          "download_url": {
                            "type": "string"
                          },
                          "format": {
                            "enum": [
                              "markdown",
                              "docx",
                              "epub"
                            ],
                            "type": "string"
                          },
                          "id": {
                            "type": "string"
                          },
                          "project_id": {
                            "type": "string"
                          },
                          "size_bytes": {
                            "minimum": 0,
                            "type": "integer"
                          },
                          "snapshot_id": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "project_id",
                          "snapshot_id",
                          "format",
                          "size_bytes",
                          "checksum_sha256",
                          "created_at",
                          "download_url"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "exports"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unified error envelope: every API failure renders as {error:{code,message,details?}}."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unified error envelope: every API failure renders as {error:{code,message,details?}}."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unified error envelope: every API failure renders as {error:{code,message,details?}}."
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      },
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "format": {
                    "enum": [
                      "markdown",
                      "docx",
                      "epub"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "format"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "created_at": {
                      "type": "string"
                    },
                    "document_id": {
                      "nullable": true,
                      "type": "string"
                    },
                    "error": {
                      "nullable": true,
                      "type": "string"
                    },
                    "events": {
                      "description": "Chronological trail (oldest first) on a single job payload; the jobs LIST endpoint is the spec-mandated newest-first surface.",
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "created_at": {
                            "type": "string"
                          },
                          "details": {
                            "additionalProperties": true,
                            "type": "object"
                          },
                          "id": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "status",
                          "details",
                          "created_at"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "id": {
                      "type": "string"
                    },
                    "kind": {
                      "type": "string"
                    },
                    "model": {
                      "type": "string"
                    },
                    "operation": {
                      "type": "string"
                    },
                    "project_id": {
                      "type": "string"
                    },
                    "provider": {
                      "type": "string"
                    },
                    "request": {
                      "additionalProperties": true,
                      "type": "object"
                    },
                    "result": {
                      "additionalProperties": true,
                      "type": "object"
                    },
                    "retry_of_job_id": {
                      "nullable": true,
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    },
                    "updated_at": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "project_id",
                    "document_id",
                    "kind",
                    "operation",
                    "status",
                    "provider",
                    "model",
                    "request",
                    "result",
                    "error",
                    "retry_of_job_id",
                    "created_at",
                    "updated_at",
                    "events"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unified error envelope: every API failure renders as {error:{code,message,details?}}."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unified error envelope: every API failure renders as {error:{code,message,details?}}."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unified error envelope: every API failure renders as {error:{code,message,details?}}."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "properties": {
                        "code": {
                          "enum": [
                            "OPERATION_IN_FLIGHT"
                          ],
                          "type": "string"
                        },
                        "details": {
                          "additionalProperties": false,
                          "properties": {
                            "document_id": {
                              "nullable": true,
                              "type": "string"
                            },
                            "operation": {
                              "type": "string"
                            },
                            "project_id": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "project_id",
                            "document_id",
                            "operation"
                          ],
                          "type": "object"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "details"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unified error envelope: every API failure renders as {error:{code,message,details?}}."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unified error envelope: every API failure renders as {error:{code,message,details?}}."
          }
        }
      }
    },
    "/api/projects/{projectId}/exports/{exportId}/download": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "exportId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/epub+zip": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              },
              "application/vnd.openxmlformats-officedocument.wordprocessingml.document": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              },
              "text/markdown; charset=utf-8": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Default Response",
            "headers": {
              "Content-Disposition": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/epub+zip": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              },
              "application/vnd.openxmlformats-officedocument.wordprocessingml.document": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              },
              "text/markdown; charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unified error envelope: every API failure renders as {error:{code,message,details?}}."
          },
          "404": {
            "content": {
              "application/epub+zip": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              },
              "application/vnd.openxmlformats-officedocument.wordprocessingml.document": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              },
              "text/markdown; charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unified error envelope: every API failure renders as {error:{code,message,details?}}."
          },
          "503": {
            "content": {
              "application/epub+zip": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              },
              "application/vnd.openxmlformats-officedocument.wordprocessingml.document": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              },
              "text/markdown; charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unified error envelope: every API failure renders as {error:{code,message,details?}}."
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
...
  }
}